requirements.txt
noobstuffs>=1.0.3
python-telegram-bot>=20.0a2
